+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_24_08_2025_13_56__Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_24_08_2025_13_56_[Changes]" date="1756033008305" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_24_08_2025_13_56_[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 24.08.2025 13:56 [Changes]" />
+</changelist>
