+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_24_08_2025_13_56__Changes_1.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_24_08_2025_13_56_[Changes]1" date="1756033034505" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_24_08_2025_13_56_[Changes]1/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 24.08.2025 13:56 [Changes]" />
+</changelist>
